requirements.txt
django==4.2
djangorestframework
django-cors-headers
django-filter
django-jazzmin
django-crum
django-environ
djangorestframework-simplejwt
djoser==2.1.0
drf-spectacular==0.25.1
django-debug-toolbar

aiogram==2.25.1
requests

black
isort
pillow
psycopg2-binary